pyproject.toml
[tool.poetry]
name = "game_api"
version = "0.1.0"
description = "API para el juego Game of Roles v2"
authors = ["Tu Nombre <tu@email.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
Flask = "3.0.2"
Flask-SQLAlchemy = "3.1.1"
Flask-Migrate = "4.0.5"
Flask-CORS = "4.0.0"
psycopg2-binary = "2.9.9"
flask-marshmallow = "1.1.0"
marshmallow-sqlalchemy = "0.30.0"
requests = "2.31.0"
pillow = "^10.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"